+++ server/src/controllers/recipeController.js
@@ -1,5 +1,6 @@
 const Recipe = require('../models/Recipe');
 const Ingredient = require('../models/Ingredient');
+const { logAction } = require('../utils/auditLog');
 
 exports.createRecipe = async (req, res) => {
   try {
@@ -15,6 +16,7 @@ exports.createRecipe = async (req, res) => {
       createdBy: req.user.id,
       version: 1
     });
+    await logAction(req, { action: 'create', entityType: 'Recipe', entityId: recipe._id, entityLabel: recipe.name });
     res.status(201).json(recipe);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -23,7 +25,7 @@ exports.createRecipe = async (req, res) => {
 
 exports.getRecipes = async (req, res) => {
   try {
-    const recipes = await Recipe.find().populate('ingredients.ingredientId', 'name unit');
+    const recipes = await Recipe.find({ isDeleted: { $ne: true } }).populate('ingredients.ingredientId', 'name unit');
     res.json(recipes);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -57,6 +59,7 @@ exports.updateRecipe = async (req, res) => {
       createdBy: req.user.id
     });
     await newRecipe.save();
+    await logAction(req, { action: 'update', entityType: 'Recipe', entityId: newRecipe._id, entityLabel: newRecipe.name, details: `New version ${version}` });
     res.json(newRecipe);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -65,8 +68,50 @@ exports.updateRecipe = async (req, res) => {
 
 exports.deleteRecipe = async (req, res) => {
   try {
+    const recipe = await Recipe.findById(req.params.id);
+    if (!recipe) return res.status(404).json({ message: 'Recipe not found' });
+    recipe.isDeleted = true;
+    recipe.deletedAt = new Date();
+    recipe.deletedBy = req.user.id;
+    await recipe.save();
+    await logAction(req, { action: 'delete', entityType: 'Recipe', entityId: recipe._id, entityLabel: recipe.name });
+    res.json({ message: 'Recipe moved to trash' });
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
+exports.getDeletedRecipes = async (req, res) => {
+  try {
+    const items = await Recipe.find({ isDeleted: true }).populate('deletedBy', 'name').sort({ deletedAt: -1 });
+    res.json(items);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
+exports.restoreRecipe = async (req, res) => {
+  try {
+    const recipe = await Recipe.findById(req.params.id);
+    if (!recipe) return res.status(404).json({ message: 'Recipe not found' });
+    recipe.isDeleted = false;
+    recipe.deletedAt = undefined;
+    recipe.deletedBy = undefined;
+    await recipe.save();
+    await logAction(req, { action: 'restore', entityType: 'Recipe', entityId: recipe._id, entityLabel: recipe.name });
+    res.json(recipe);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
+exports.permanentlyDeleteRecipe = async (req, res) => {
+  try {
+    const recipe = await Recipe.findById(req.params.id);
+    if (!recipe) return res.status(404).json({ message: 'Recipe not found' });
+    await logAction(req, { action: 'permanent_delete', entityType: 'Recipe', entityId: recipe._id, entityLabel: recipe.name });
     await Recipe.findByIdAndDelete(req.params.id);
-    res.json({ message: 'Recipe deleted' });
+    res.json({ message: 'Recipe permanently deleted' });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -84,6 +129,7 @@ exports.duplicateRecipe = async (req, res) => {
       version: 1
     });
     await newRecipe.save();
+    await logAction(req, { action: 'create', entityType: 'Recipe', entityId: newRecipe._id, entityLabel: newRecipe.name, details: `Duplicated from ${recipe.name}` });
     res.status(201).json(newRecipe);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -96,6 +142,7 @@ exports.toggleRecipeStatus = async (req, res) => {
     if (!recipe) return res.status(404).json({ message: 'Recipe not found' });
     recipe.isActive = !recipe.isActive;
     await recipe.save();
+    await logAction(req, { action: 'update', entityType: 'Recipe', entityId: recipe._id, entityLabel: recipe.name, details: recipe.isActive ? 'Activated' : 'Deactivated' });
     res.json(recipe);
   } catch (error) {
     res.status(500).json({ message: error.message });
